Update everything
This updates the Rust edition to 2024, all dependencies, and fixes style
issues.

With the rusqlite update (v0.38.0), we can no longer pass u64/usize into
queries, so some handling around content length has been changed to
explicitly use i32, just as Bytebin does.

+++ src/db.rs
@@ -1,6 +1,6 @@
 use actix_web::web;
 use anyhow::Result;
-use rusqlite::{types::Null, OptionalExtension};
+use rusqlite::{OptionalExtension, types::Null};
 use serde::{Deserialize, Serialize};
 
 pub type Pool = r2d2::Pool<r2d2_sqlite::SqliteConnectionManager>;
@@ -16,7 +16,7 @@ pub struct Content {
     pub auth_key: Option<String>,
     pub content_encoding: String,
     pub backend_id: String,
-    pub content_length: usize,
+    pub content_length: i32,
     pub content: Option<Vec<u8>>,
 }
